requirements/base.txt
python-slugify>=8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi>=23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
redis>=5.0.7  # https://github.com/redis/redis-py
hiredis>=2.3.2  # https://github.com/redis/hiredis-py
celery>=5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat>=2.6.0  # https://github.com/celery/django-celery-beat
flower>=2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django>=5.0.6  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ>=0.11.2  # https://github.com/joke2k/django-environ
django-model-utils>=4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==0.59.0 # https://github.com/pennersr/django-allauth
django-crispy-forms>=2.2  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-tailwind  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis>=5.4.0  # https://github.com/jazzband/django-redis
django-webpack-loader>=3.1.0  # https://github.com/django-webpack/django-webpack-loader
django-filter>=24.2
# django-simple-captcha>=0.6.0
django-recaptcha==4.0.0
django-htmx>=1.17.3
django-widget-tweaks>=1.5.0
django-tables2>=2.7.0
django-render-block>=0.9.2
Scrapy
scrapy-redis>=0.7.3
celery-progress>=0.3
celery-progress[websockets,redis]
dateparser>=1.2.0
python-decouple>=3.8
django-celery-results>=2.5.1
tablib>=3.6.1
django_components>=0.81
django-mptt>=0.16.0
django-ckeditor
django-ckeditor-5
black>=24.4.2
qrcode>=7.0.0
olefile>=0.47
